+++ service-task/src/docs/asciidoc/manual_content2.adoc
@@ -1,0 +1,4 @@
+== Chapter of manual content 2
+
+This is some dummy text
+
