--- service-task/src/main/java/com/github/hollykunge/security/constants/TaskContants.java
@@ -1,22 +1,0 @@
-package com.github.hollykunge.security.constants;
-
-/**
- * @author: zhhongyu
- * @description: 常量类
- * @since: Create in 9:56 2019/9/10
- */
-public class TaskContants {
-    public static final String GIT_REF_HEAD = "refs/heads/";
-    /**
-     * merge到分支标识
-     */
-    public static final String GIT_BRANCH_TO_MERGE_BRANCH = "to";
-    /**
-     * 从哪个分支合并标识
-     */
-    public static final String GIT_BRANCH_FROM_MERGE_BRANCH = "from";
-    /**
-     * git最牛逼的分支，业务不使用这个牛逼分支
-     */
-    public static final String GIT_SURPER_BRANCH = "master";
-}
