--- service-task/src/main/java/com/github/hollykunge/security/mapper/TaskInfoMapper.java
@@ -1,12 +1,0 @@
-package com.github.hollykunge.security.mapper;
-
-import com.github.hollykunge.security.entity.TaskEntity;
-import tk.mybatis.mapper.common.Mapper;
-
-/**
- * @author: zhhongyu
- * @description: 任务do接口
- * @since: Create in 11:05 2019/9/5
- */
-public interface TaskInfoMapper extends Mapper<TaskEntity> {
-}
