--- service-task/src/main/java/com/github/hollykunge/security/mapper/UserTaskMapMapper.java
@@ -1,12 +1,0 @@
-package com.github.hollykunge.security.mapper;
-
-import com.github.hollykunge.security.entity.UserTaskMap;
-import tk.mybatis.mapper.common.Mapper;
-
-/**
- * @author: zhhongyu
- * @description: 用户任务do接口
- * @since: Create in 11:05 2019/9/5
- */
-public interface UserTaskMapMapper extends Mapper<UserTaskMap> {
-}
