--- service-task/src/main/java/com/github/hollykunge/security/util/ChineseUtils.java
@@ -1,60 +1,0 @@
-package com.github.hollykunge.security.util;
-
-import net.sourceforge.pinyin4j.PinyinHelper;
-import net.sourceforge.pinyin4j.format.HanyuPinyinCaseType;
-import net.sourceforge.pinyin4j.format.HanyuPinyinOutputFormat;
-import net.sourceforge.pinyin4j.format.HanyuPinyinToneType;
-import net.sourceforge.pinyin4j.format.exception.BadHanyuPinyinOutputFormatCombination;
-
-/**
- * @author: zhhongyu
- * @description:
- * @since: Create in 10:19 2019/9/23
- */
-public class ChineseUtils {
-    public static String toPinyin(String chinese) {
-        String pinyinStr = "";
-        char[] newChar = chinese.toCharArray();
-        HanyuPinyinOutputFormat defaultFormat = new HanyuPinyinOutputFormat();
-        defaultFormat.setCaseType(HanyuPinyinCaseType.LOWERCASE);
-        defaultFormat.setToneType(HanyuPinyinToneType.WITHOUT_TONE);
-        for (int i = 0; i < newChar.length; i++) {
-            if (newChar[i] > 128) {
-                try {
-                    pinyinStr += PinyinHelper.toHanyuPinyinStringArray(newChar[i], defaultFormat)[0];
-                } catch (BadHanyuPinyinOutputFormatCombination e) {
-                    e.printStackTrace();
-                }
-            } else {
-                pinyinStr += newChar[i];
-            }
-        }
-        return pinyinStr;
-    }
-    public static String toFirstChar(String chinese){
-        String pinyinStr = "";
-        char[] newChar = chinese.toCharArray();  //转为单个字符
-        HanyuPinyinOutputFormat defaultFormat = new HanyuPinyinOutputFormat();
-        defaultFormat.setCaseType(HanyuPinyinCaseType.LOWERCASE);
-        defaultFormat.setToneType(HanyuPinyinToneType.WITHOUT_TONE);
-        for (int i = 0; i < newChar.length; i++) {
-            if (newChar[i] > 128) {
-                try {
-                    pinyinStr += PinyinHelper.toHanyuPinyinStringArray(newChar[i], defaultFormat)[0].charAt(0);
-                } catch (BadHanyuPinyinOutputFormatCombination e) {
-                    e.printStackTrace();
-                }
-            }else{
-                pinyinStr += newChar[i];
-            }
-        }
-        return pinyinStr;
-    }
-
-    public static void main(String[] args) {
-        String aa = ChineseUtils.toPinyin("中国好声音");
-        String bb = ChineseUtils.toFirstChar("lark_中国好声音");
-        System.out.println(aa);
-        System.out.println(bb);
-    }
-}
