+++ service-task/src/main/resources/application-dev.yml
@@ -1,0 +1,15 @@
+#spring:
+#  datasource:
+#    url: jdbc:h2:mem:task_db;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+#    username: root
+#    password: 123456
+#  h2:
+#    console:
+#      enabled: true
+#      path: /h2-console
+spring:
+  data:
+    mongodb:
+      host: localhost
+      port: 27017
+      database: task
