--- service-task/src/main/resources/mapper/UserTaskMapMapper.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.github.hollykunge.security.mapper.UserTaskMapMapper">
-    <resultMap id="BaseResultMap" type="com.github.hollykunge.security.entity.UserTaskMap">
-        <id column="ID" property="id" jdbcType="VARCHAR"/>
-        <result column="USER_ID" property="userId" jdbcType="VARCHAR"/>
-        <result column="TASK_ID" property="taskId" jdbcType="VARCHAR"/>
-        <result column="USER_NAME" property="userName" jdbcType="VARCHAR"/>
-        <result column="TASK_NAME" property="taskName" jdbcType="VARCHAR"/>
-        <result column="PERMISSION" property="permission" jdbcType="TIMESTAMP"/>
-        <result column="CRT_TIME" property="crtTime" jdbcType="TIMESTAMP"/>
-        <result column="CRT_USER" property="crtUser" jdbcType="VARCHAR"/>
-        <result column="CRT_NAME" property="crtName" jdbcType="VARCHAR"/>
-        <result column="CRT_HOST" property="crtHost" jdbcType="VARCHAR"/>
-        <result column="ATTR1" property="attr1" jdbcType="VARCHAR"/>
-        <result column="ATTR2" property="attr2" jdbcType="VARCHAR"/>
-        <result column="ATTR3" property="attr3" jdbcType="VARCHAR"/>
-        <result column="ATTR4" property="attr4" jdbcType="VARCHAR"/>
-    </resultMap>
-</mapper>
